requirements.txt
Flask==2.2.5
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0
python-jose==3.3.0
google-cloud-datastore==2.17.0
google-cloud-storage==2.14.0
